rm VecAddKernel's copy
